moved things to the right spots
