+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/IntoTheDeep24_25/teleop/arm/teleop_arm_code.java
@@ -1,11 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.IntoTheDeep24_25.teleop.arm;
 
 import static com.qualcomm.robotcore.util.Range.clip;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
 
 import org.firstinspires.ftc.teamcode.PowerPlay23_24.TemplateJanx;
